dashboard/requirements.txt
flask==3.0.0
docker==7.0.0
requests==2.31.0
semver==3.0.2
gunicorn==21.2.0
python-dotenv==1.0.0

